fix(bridgeGame): 요구사항에 따라 BridgeGame 으로부터 InputView와 OutputView 분리
